Update to MediaWiki 1.32.0

+++ includes/changetags/ChangeTags.php
@@ -22,6 +22,7 @@
  */
 
 use MediaWiki\MediaWikiServices;
+use MediaWiki\Storage\NameTableAccessException;
 use Wikimedia\Rdbms\Database;
 
 class ChangeTags {
@@ -32,6 +33,9 @@ class ChangeTags {
 	 */
 	const MAX_DELETE_USES = 5000;
 
+	/**
+	 * A list of tags defined and used by MediaWiki itself.
+	 */
 	private static $definedSoftwareTags = [
 		'mw-contentmodelchange',
 		'mw-new-redirect',
@@ -84,6 +88,7 @@ class ChangeTags {
 	 * @return array Array with two items: (html, classes)
 	 *   - html: String: HTML for displaying the tags (empty string when param $tags is empty)
 	 *   - classes: Array of strings: CSS classes used in the generated html, one class for each tag
+	 * @return-taint onlysafefor_htmlnoent
 	 */
 	public static function formatSummaryRow( $tags, $page, IContextSource $context = null ) {
 		if ( !$tags ) {
@@ -198,9 +203,8 @@ class ChangeTags {
 		}
 
 		$taglessDesc = Sanitizer::stripAllTags( $originalDesc->parse() );
-		$escapedDesc = Sanitizer::escapeHtmlAllowEntities( $taglessDesc );
 
-		return $context->getLanguage()->truncateForVisual( $escapedDesc, $length );
+		return $context->getLanguage()->truncateForVisual( $taglessDesc, $length );
 	}
 
 	/**
@@ -210,7 +214,7 @@ class ChangeTags {
 	 * @param int|null $rc_id The rc_id of the change to add the tags to
 	 * @param int|null $rev_id The rev_id of the change to add the tags to
 	 * @param int|null $log_id The log_id of the change to add the tags to
-	 * @param string $params Params to put in the ct_params field of table 'change_tag'
+	 * @param string|null $params Params to put in the ct_params field of table 'change_tag'
 	 * @param RecentChange|null $rc Recent change, in case the tagging accompanies the action
 	 * (this should normally be the case)
 	 *
@@ -241,7 +245,7 @@ class ChangeTags {
 	 * Pass a variable whose value is null if the rev_id is not relevant or unknown.
 	 * @param int|null &$log_id The log_id of the change to add the tags to.
 	 * Pass a variable whose value is null if the log_id is not relevant or unknown.
-	 * @param string $params Params to put in the ct_params field of table
+	 * @param string|null $params Params to put in the ct_params field of table
 	 * 'change_tag' when adding tags
 	 * @param RecentChange|null $rc Recent change being tagged, in case the tagging accompanies
 	 * the action
@@ -258,6 +262,8 @@ class ChangeTags {
 		&$rev_id = null, &$log_id = null, $params = null, RecentChange $rc = null,
 		User $user = null
 	) {
+		global $wgChangeTagsSchemaMigrationStage;
+
 		$tagsToAdd = array_filter( (array)$tagsToAdd ); // Make sure we're submitting all tags...
 		$tagsToRemove = array_filter( (array)$tagsToRemove );
 
@@ -338,22 +344,44 @@ class ChangeTags {
 		}
 
 		// insert a row into change_tag for each new tag
+		$changeTagDefStore = MediaWikiServices::getInstance()->getChangeTagDefStore();
 		if ( count( $tagsToAdd ) ) {
+			$changeTagMapping = [];
+			if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD ) {
+				foreach ( $tagsToAdd as $tag ) {
+					$changeTagMapping[$tag] = $changeTagDefStore->acquireId( $tag );
+				}
+
+				$dbw->update(
+					'change_tag_def',
+					[ 'ctd_count = ctd_count + 1' ],
+					[ 'ctd_name' => $tagsToAdd ],
+					__METHOD__
+				);
+			}
+
 			$tagsRows = [];
 			foreach ( $tagsToAdd as $tag ) {
+				if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+					$tagName = null;
+				} else {
+					$tagName = $tag;
+				}
 				// Filter so we don't insert NULLs as zero accidentally.
 				// Keep in mind that $rc_id === null means "I don't care/know about the
 				// rc_id, just delete $tag on this revision/log entry". It doesn't
 				// mean "only delete tags on this revision/log WHERE rc_id IS NULL".
 				$tagsRows[] = array_filter(
 					[
-						'ct_tag' => $tag,
+						'ct_tag' => $tagName,
 						'ct_rc_id' => $rc_id,
 						'ct_log_id' => $log_id,
 						'ct_rev_id' => $rev_id,
-						'ct_params' => $params
+						'ct_params' => $params,
+						'ct_tag_id' => $changeTagMapping[$tag] ?? null,
 					]
 				);
+
 			}
 
 			$dbw->insert( 'change_tag', $tagsRows, __METHOD__, [ 'IGNORE' ] );
@@ -362,15 +390,37 @@ class ChangeTags {
 		// delete from change_tag
 		if ( count( $tagsToRemove ) ) {
 			foreach ( $tagsToRemove as $tag ) {
+				if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+					$tagName = null;
+					$tagId = $changeTagDefStore->getId( $tag );
+				} else {
+					$tagName = $tag;
+					$tagId = null;
+				}
 				$conds = array_filter(
 					[
-						'ct_tag' => $tag,
+						'ct_tag' => $tagName,
 						'ct_rc_id' => $rc_id,
 						'ct_log_id' => $log_id,
-						'ct_rev_id' => $rev_id
+						'ct_rev_id' => $rev_id,
+						'ct_tag_id' => $tagId,
 					]
 				);
 				$dbw->delete( 'change_tag', $conds, __METHOD__ );
+				if ( $dbw->affectedRows() && $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD ) {
+					$dbw->update(
+						'change_tag_def',
+						[ 'ctd_count = ctd_count - 1' ],
+						[ 'ctd_name' => $tag ],
+						__METHOD__
+					);
+
+					$dbw->delete(
+						'change_tag_def',
+						[ 'ctd_name' => $tag, 'ctd_count' => 0, 'ctd_user_defined' => 0 ],
+						__METHOD__
+					);
+				}
 			}
 		}
 
@@ -416,7 +466,7 @@ class ChangeTags {
 		// $prevTags can be out of date on replica DBs, especially when addTags is called consecutively,
 		// causing loss of tags added recently in tag_summary table.
 		$prevTags = $dbw->selectField( 'tag_summary', 'ts_tags', $tsConds, __METHOD__ );
-		$prevTags = $prevTags ? $prevTags : '';
+		$prevTags = $prevTags ?: '';
 		$prevTags = array_filter( explode( ',', $prevTags ) );
 
 		// add tags
@@ -474,9 +524,12 @@ class ChangeTags {
 	 * Is it OK to allow the user to apply all the specified tags at the same time
 	 * as they edit/make the change?
 	 *
+	 * Extensions should not use this function, unless directly handling a user
+	 * request to add a tag to a revision or log entry that the user is making.
+	 *
 	 * @param array $tags Tags that you are interested in applying
-	 * @param User|null $user User whose permission you wish to check, or null if
-	 * you don't care (e.g. maintenance scripts)
+	 * @param User|null $user User whose permission you wish to check, or null to
+	 * check for a generic non-blocked user with the relevant rights
 	 * @return Status
 	 * @since 1.25
 	 */
@@ -541,10 +594,13 @@ class ChangeTags {
 	 * Is it OK to allow the user to adds and remove the given tags tags to/from a
 	 * change?
 	 *
+	 * Extensions should not use this function, unless directly handling a user
+	 * request to add or remove tags from an existing revision or log entry.
+	 *
 	 * @param array $tagsToAdd Tags that you are interested in adding
 	 * @param array $tagsToRemove Tags that you are interested in removing
-	 * @param User|null $user User whose permission you wish to check, or null if
-	 * you don't care (e.g. maintenance scripts)
+	 * @param User|null $user User whose permission you wish to check, or null to
+	 * check for a generic non-blocked user with the relevant rights
 	 * @return Status
 	 * @since 1.25
 	 */
@@ -589,10 +645,14 @@ class ChangeTags {
 	 * Adds and/or removes tags to/from a given change, checking whether it is
 	 * allowed first, and adding a log entry afterwards.
 	 *
-	 * Includes a call to ChangeTag::canUpdateTags(), so your code doesn't need
+	 * Includes a call to ChangeTags::canUpdateTags(), so your code doesn't need
 	 * to do that. However, it doesn't check whether the *_id parameters are a
 	 * valid combination. That is up to you to enforce. See ApiTag::execute() for
 	 * an example.
+	 *
+	 * Extensions should generally avoid this function. Call
+	 * ChangeTags::updateTags() instead, unless directly handling a user request
+	 * to add or remove tags from an existing revision or log entry.
 	 *
 	 * @param array|null $tagsToAdd If none, pass array() or null
 	 * @param array|null $tagsToRemove If none, pass array() or null
@@ -721,8 +781,9 @@ class ChangeTags {
 	 * @throws MWException When unable to determine appropriate JOIN condition for tagging
 	 */
 	public static function modifyDisplayQuery( &$tables, &$fields, &$conds,
-										&$join_conds, &$options, $filter_tag = '' ) {
-		global $wgUseTagFilter;
+		&$join_conds, &$options, $filter_tag = ''
+	) {
+		global $wgUseTagFilter, $wgChangeTagsSchemaMigrationStage;
 
 		// Normalize to arrays
 		$tables = (array)$tables;
@@ -743,8 +804,18 @@ class ChangeTags {
 			throw new MWException( 'Unable to determine appropriate JOIN condition for tagging.' );
 		}
 
+		$tagTables[] = 'change_tag';
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+			$tagTables[] = 'change_tag_def';
+			$join_cond_ts_tags = [ $join_cond, 'ct_tag_id=ctd_id' ];
+			$field = 'ctd_name';
+		} else {
+			$field = 'ct_tag';
+			$join_cond_ts_tags = $join_cond;
+		}
+
 		$fields['ts_tags'] = wfGetDB( DB_REPLICA )->buildGroupConcatField(
-			',', 'change_tag', 'ct_tag', $join_cond
+			',', $tagTables, $field, $join_cond_ts_tags
 		);
 
 		if ( $wgUseTagFilter && $filter_tag ) {
@@ -753,7 +824,26 @@ class ChangeTags {
 
 			$tables[] = 'change_tag';
 			$join_conds['change_tag'] = [ 'INNER JOIN', $join_cond ];
-			$conds['ct_tag'] = $filter_tag;
+			if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+				$filterTagIds = [];
+				$changeTagDefStore = MediaWikiServices::getInstance()->getChangeTagDefStore();
+				foreach ( (array)$filter_tag as $filterTagName ) {
+					try {
+						$filterTagIds[] = $changeTagDefStore->getId( $filterTagName );
+					} catch ( NameTableAccessException $exception ) {
+						// Return nothing.
+						$conds[] = '0';
+						break;
+					};
+				}
+
+				if ( $filterTagIds !== [] ) {
+					$conds['ct_tag_id'] = $filterTagIds;
+				}
+			} else {
+				$conds['ct_tag'] = $filter_tag;
+			}
+
 			if (
 				is_array( $filter_tag ) && count( $filter_tag ) > 1 &&
 				!in_array( 'DISTINCT', $options )
@@ -814,8 +904,8 @@ class ChangeTags {
 	}
 
 	/**
-	 * Defines a tag in the valid_tag table, without checking that the tag name
-	 * is valid.
+	 * Defines a tag in the valid_tag table and/or update ctd_user_defined field in change_tag_def,
+	 * without checking that the tag name is valid.
 	 * Extensions should NOT use this function; they can use the ListDefinedTags
 	 * hook instead.
 	 *
@@ -823,27 +913,67 @@ class ChangeTags {
 	 * @since 1.25
 	 */
 	public static function defineTag( $tag ) {
-		$dbw = wfGetDB( DB_MASTER );
-		$dbw->replace( 'valid_tag',
-			[ 'vt_tag' ],
-			[ 'vt_tag' => $tag ],
-			__METHOD__ );
+		global $wgChangeTagsSchemaMigrationStage;
 
+		$dbw = wfGetDB( DB_MASTER );
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD ) {
+			$tagDef = [
+				'ctd_name' => $tag,
+				'ctd_user_defined' => 1,
+				'ctd_count' => 0
+			];
+			$dbw->upsert(
+				'change_tag_def',
+				$tagDef,
+				[ 'ctd_name' ],
+				[ 'ctd_user_defined' => 1 ],
+				__METHOD__
+			);
+		}
+
+		if ( $wgChangeTagsSchemaMigrationStage < MIGRATION_NEW ) {
+			$dbw->replace(
+				'valid_tag',
+				[ 'vt_tag' ],
+				[ 'vt_tag' => $tag ],
+				__METHOD__
+			);
+		}
 		// clear the memcache of defined tags
 		self::purgeTagCacheAll();
 	}
 
 	/**
-	 * Removes a tag from the valid_tag table. The tag may remain in use by
-	 * extensions, and may still show up as 'defined' if an extension is setting
-	 * it from the ListDefinedTags hook.
+	 * Removes a tag from the valid_tag table and/or update ctd_user_defined field in change_tag_def.
+	 * The tag may remain in use by extensions, and may still show up as 'defined'
+	 * if an extension is setting it from the ListDefinedTags hook.
 	 *
 	 * @param string $tag Tag to remove
 	 * @since 1.25
 	 */
 	public static function undefineTag( $tag ) {
+		global $wgChangeTagsSchemaMigrationStage;
+
 		$dbw = wfGetDB( DB_MASTER );
-		$dbw->delete( 'valid_tag', [ 'vt_tag' => $tag ], __METHOD__ );
+
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD ) {
+			$dbw->update(
+				'change_tag_def',
+				[ 'ctd_user_defined' => 0 ],
+				[ 'ctd_name' => $tag ],
+				__METHOD__
+			);
+
+			$dbw->delete(
+				'change_tag_def',
+				[ 'ctd_name' => $tag, 'ctd_count' => 0 ],
+				__METHOD__
+			);
+		}
+
+		if ( $wgChangeTagsSchemaMigrationStage < MIGRATION_NEW ) {
+			$dbw->delete( 'valid_tag', [ 'vt_tag' => $tag ], __METHOD__ );
+		}
 
 		// clear the memcache of defined tags
 		self::purgeTagCacheAll();
@@ -856,7 +986,7 @@ class ChangeTags {
 	 * @param string $tag
 	 * @param string $reason
 	 * @param User $user Who to attribute the action to
-	 * @param int $tagCount For deletion only, how many usages the tag had before
+	 * @param int|null $tagCount For deletion only, how many usages the tag had before
 	 * it was deleted.
 	 * @param array $logEntryTags Change tags to apply to the entry
 	 * that will be created in the tag management log
@@ -1057,6 +1187,9 @@ class ChangeTags {
 	/**
 	 * Is it OK to allow the user to create this tag?
 	 *
+	 * Extensions should NOT use this function. In most cases, a tag can be
+	 * defined using the ListDefinedTags hook without any checking.
+	 *
 	 * @param string $tag Tag that you are interested in creating
 	 * @param User|null $user User whose permission you wish to check, or null if
 	 * you don't care (e.g. maintenance scripts)
@@ -1091,8 +1224,12 @@ class ChangeTags {
 
 	/**
 	 * Creates a tag by adding a row to the `valid_tag` table.
+	 * and/or add it to `change_tag_def` table.
 	 *
-	 * Includes a call to ChangeTag::canDeleteTag(), so your code doesn't need to
+	 * Extensions should NOT use this function; they can use the ListDefinedTags
+	 * hook instead.
+	 *
+	 * Includes a call to ChangeTag::canCreateTag(), so your code doesn't need to
 	 * do that.
 	 *
 	 * @param string $tag
@@ -1138,16 +1275,24 @@ class ChangeTags {
 	 * @since 1.25
 	 */
 	public static function deleteTagEverywhere( $tag ) {
+		global $wgChangeTagsSchemaMigrationStage;
 		$dbw = wfGetDB( DB_MASTER );
 		$dbw->startAtomic( __METHOD__ );
 
-		// delete from valid_tag
+		// delete from valid_tag and/or set ctd_user_defined = 0
 		self::undefineTag( $tag );
+
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+			$tagId = MediaWikiServices::getInstance()->getChangeTagDefStore()->getId( $tag );
+			$conditions = [ 'ct_tag_id' => $tagId ];
+		} else {
+			$conditions = [ 'ct_tag' => $tag ];
+		}
 
 		// find out which revisions use this tag, so we can delete from tag_summary
 		$result = $dbw->select( 'change_tag',
-			[ 'ct_rc_id', 'ct_log_id', 'ct_rev_id', 'ct_tag' ],
-			[ 'ct_tag' => $tag ],
+			[ 'ct_rc_id', 'ct_log_id', 'ct_rev_id' ],
+			$conditions,
 			__METHOD__ );
 		foreach ( $result as $row ) {
 			// remove the tag from the relevant row of tag_summary
@@ -1158,7 +1303,16 @@ class ChangeTags {
 		}
 
 		// delete from change_tag
-		$dbw->delete( 'change_tag', [ 'ct_tag' => $tag ], __METHOD__ );
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+			$tagId = MediaWikiServices::getInstance()->getChangeTagDefStore()->getId( $tag );
+			$dbw->delete( 'change_tag', [ 'ct_tag_id' => $tagId ], __METHOD__ );
+		} else {
+			$dbw->delete( 'change_tag', [ 'ct_tag' => $tag ], __METHOD__ );
+		}
+
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD ) {
+			$dbw->delete( 'change_tag_def', [ 'ctd_name' => $tag ], __METHOD__ );
+		}
 
 		$dbw->endAtomic( __METHOD__ );
 
@@ -1248,7 +1402,7 @@ class ChangeTags {
 
 		// store the tag usage statistics
 		$tagUsage = self::tagUsageStatistics();
-		$hitcount = isset( $tagUsage[$tag] ) ? $tagUsage[$tag] : 0;
+		$hitcount = $tagUsage[$tag] ?? 0;
 
 		// do it!
 		$deleteResult = self::deleteTagEverywhere( $tag );
@@ -1310,7 +1464,7 @@ class ChangeTags {
 	/**
 	 * Lists tags explicitly defined in the `valid_tag` table of the database.
 	 * Tags in table 'change_tag' which are not in table 'valid_tag' are not
-	 * included.
+	 * included. In case of new backend loads the data from `change_tag_def` table.
 	 *
 	 * Tries memcached first.
 	 *
@@ -1325,11 +1479,16 @@ class ChangeTags {
 			$cache->makeKey( 'valid-tags-db' ),
 			WANObjectCache::TTL_MINUTE * 5,
 			function ( $oldValue, &$ttl, array &$setOpts ) use ( $fname ) {
+				global $wgChangeTagsSchemaMigrationStage;
 				$dbr = wfGetDB( DB_REPLICA );
 
 				$setOpts += Database::getCacheSetOptions( $dbr );
 
-				$tags = $dbr->selectFieldValues( 'valid_tag', 'vt_tag', [], $fname );
+				if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ) {
+					$tags = self::listExplicitlyDefinedTagsNewBackend();
+				} else {
+					$tags = $dbr->selectFieldValues( 'valid_tag', 'vt_tag', [], $fname );
+				}
 
 				return array_filter( array_unique( $tags ) );
 			},
@@ -1338,6 +1497,22 @@ class ChangeTags {
 				'lockTSE' => WANObjectCache::TTL_MINUTE * 5,
 				'pcTTL' => WANObjectCache::TTL_PROC_LONG
 			]
+		);
+	}
+
+	/**
+	 * Lists tags explicitly user defined tags. When ctd_user_defined is true.
+	 *
+	 * @return string[] Array of strings: tags
+	 * @since 1.25
+	 */
+	private static function listExplicitlyDefinedTagsNewBackend() {
+		$dbr = wfGetDB( DB_REPLICA );
+		return $dbr->selectFieldValues(
+			'change_tag_def',
+			'ctd_name',
+			[ 'ctd_user_defined' => 1 ],
+			__METHOD__
 		);
 	}
 
@@ -1410,6 +1585,13 @@ class ChangeTags {
 	 * @return array Array of string => int
 	 */
 	public static function tagUsageStatistics() {
+		global $wgChangeTagsSchemaMigrationStage, $wgTagStatisticsNewTable;
+		if ( $wgChangeTagsSchemaMigrationStage > MIGRATION_WRITE_BOTH ||
+			( $wgTagStatisticsNewTable && $wgChangeTagsSchemaMigrationStage > MIGRATION_OLD )
+		) {
+			return self::newTagUsageStatistics();
+		}
+
 		$fname = __METHOD__;
 		$cache = MediaWikiServices::getInstance()->getMainWANObjectCache();
 		return $cache->getWithSetCallback(
@@ -1441,6 +1623,29 @@ class ChangeTags {
 				'pcTTL' => WANObjectCache::TTL_PROC_LONG
 			]
 		);
+	}
+
+	/**
+	 * Same self::tagUsageStatistics() but uses change_tag_def.
+	 *
+	 * @return array Array of string => int
+	 */
+	private static function newTagUsageStatistics() {
+		$dbr = wfGetDB( DB_REPLICA );
+		$res = $dbr->select(
+			'change_tag_def',
+			[ 'ctd_name', 'ctd_count' ],
+			[],
+			__METHOD__,
+			[ 'ORDER BY' => 'ctd_count DESC' ]
+		);
+
+		$out = [];
+		foreach ( $res as $row ) {
+			$out[$row->ctd_name] = $row->ctd_count;
+		}
+
+		return $out;
 	}
 
 	/**
